Revert "JPKO Update"

+++ KOF.Core/Models/PartyMember.cs
@@ -6,7 +6,7 @@ namespace KOF.Core.Models;
 public class PartyMember
 {
     [Browsable(false)]
-    public short MemberId { get; set; }
+    public int MemberId { get; set; }
     [Browsable(false)]
     public byte Index { get; set; }
     public string Name { get; set; } = "";
@@ -50,7 +50,7 @@ public class PartyMember
     public static PartyMember FromMessage(Message msg) => new()
     {
         Unknown1 = msg.Read<short>(),
-        MemberId = msg.Read<short>(),
+        MemberId = msg.Read<int>(),
         Index = msg.Read<byte>(),
         Name = msg.Read(true, "gb2312"),
         MaxHealth = msg.Read<ushort>(),
